requirements.txt
browserstack-local
jsonmerge
multiprocess
paver
selenium==4.*
psutil
pytest
pytest-variables
pytest-selenium
pytest-xdist